remove ffmpge
